Wire monthly public, homepage and admin routes while preserving the final weekly round
